+++ Sources/MacAotoKill/StatusMenuController.swift
@@ -451,6 +451,7 @@ final class StatusMenuController: NSObject, NSMenuDelegate {
 
     @objc private func addWhitelistItem(_ sender: NSMenuItem) {
         guard let bundleID = sender.representedObject as? String else { return }
+        settingsStore.setMinimumBackgroundDuration(nil, for: bundleID)
         whitelistStore.add(bundleID)
         eventLog.append(localizer.t("event.addedWhitelist", bundleID))
         refreshSnapshot(performAutomaticRelease: false)